Revert "move validate details method to TseService"
This reverts commit d7a3fea5057c9156969511a11e4e888d007f622f.
