+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/RespondentTellSomethingElseService.java
@@ -13,7 +13,9 @@ import uk.gov.hmcts.et.common.model.ccd.CaseData;
 import uk.gov.hmcts.et.common.model.ccd.CaseDetails;
 import uk.gov.hmcts.et.common.model.ccd.items.GenericTseApplicationType;
 import uk.gov.hmcts.et.common.model.ccd.items.GenericTseApplicationTypeItem;
+import uk.gov.hmcts.ethos.replacement.docmosis.helpers.RespondentTellSomethingElseHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.TseViewApplicationHelper;
+import uk.gov.hmcts.ethos.replacement.docmosis.utils.RespondentTSEApplicationTypeData;
 import uk.gov.service.notify.NotificationClient;
 import uk.gov.service.notify.NotificationClientException;
 
@@ -80,6 +82,7 @@ public class RespondentTellSomethingElseService {
     private String cyTseRespondentToClaimantTemplateId;
     @Value("${template.tse.respondent.application.tribunal}")
     private String tseNewApplicationAdminTemplateId;
+    private static final String GIVE_DETAIL_MISSING = "Use the text box or file upload to give details.";
     private static final List<String> GROUP_B_TYPES = List.of(TSE_APP_CHANGE_PERSONAL_DETAILS,
             TSE_APP_CONSIDER_A_DECISION_AFRESH, TSE_APP_RECONSIDER_JUDGEMENT);
     private static final String DOCGEN_ERROR = "Failed to generate document for case id: %s";
@@ -104,6 +107,22 @@ public class RespondentTellSomethingElseService {
                     + "%s\r\n";
 
     private static final String TABLE_ROW_MARKDOWN = "|%s|%s|%s|%s|%s|%s|%s|\r\n";
+
+    /**
+     * Validate Give Details (free text box) or file upload is mandatory.
+     *
+     * @param caseData in which the case details are extracted from
+     * @return errors Error message
+     */
+    public List<String> validateGiveDetails(CaseData caseData) {
+        List<String> errors = new ArrayList<>();
+        RespondentTSEApplicationTypeData selectedAppData =
+                RespondentTellSomethingElseHelper.getSelectedApplicationType(caseData);
+        if (selectedAppData.getResTseDocument() == null && isNullOrEmpty(selectedAppData.getSelectedTextBox())) {
+            errors.add(GIVE_DETAIL_MISSING);
+        }
+        return errors;
+    }
 
     /**
      * Uses {@link EmailService} to generate an email to Respondent.
