+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/service/TseServiceTest.java
@@ -1,13 +1,9 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.service;
 
-import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.et.common.model.bulk.types.DynamicFixedListType;
@@ -21,20 +17,16 @@ import uk.gov.hmcts.et.common.model.ccd.items.TseRespondTypeItem;
 import uk.gov.hmcts.et.common.model.ccd.types.DocumentType;
 import uk.gov.hmcts.et.common.model.ccd.types.TseAdminRecordDecisionType;
 import uk.gov.hmcts.et.common.model.ccd.types.TseRespondType;
-import uk.gov.hmcts.et.common.model.ccd.types.UploadedDocumentType;
 import uk.gov.hmcts.et.common.model.ccd.types.citizenhub.ClaimantTse;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.DocumentFixtures;
-import uk.gov.hmcts.ethos.utils.CaseDataBuilder;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Stream;
 
 import static java.util.Arrays.deepEquals;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.hamcrest.core.Is.is;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -44,17 +36,6 @@ import static uk.gov.hmcts.ecm.common.model.helper.Constants.CLAIMANT_TITLE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.NO;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.RESPONDENT_TITLE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_AMEND_RESPONSE;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_CHANGE_PERSONAL_DETAILS;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_CLAIMANT_NOT_COMPLIED;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_CONSIDER_A_DECISION_AFRESH;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_CONTACT_THE_TRIBUNAL;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_ORDER_A_WITNESS_TO_ATTEND_TO_GIVE_EVIDENCE;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_ORDER_OTHER_PARTY;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_POSTPONE_A_HEARING;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_RECONSIDER_JUDGEMENT;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_RESTRICT_PUBLICITY;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_STRIKE_OUT_ALL_OR_PART_OF_A_CLAIM;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_VARY_OR_REVOKE_AN_ORDER;
 
 @ExtendWith(SpringExtension.class)
 class TseServiceTest {
@@ -66,8 +47,6 @@ class TseServiceTest {
     private static final String AUTH_TOKEN = "Bearer authToken";
     private static final String COPY_CORRESPONDENCE = "Do you want to copy this correspondence to the other party to "
         + "satisfy the Rules of Procedure?";
-
-    private static final String GIVE_DETAIL_MISSING = "Use the text box or file upload to give details.";
 
     @BeforeEach
     void setUp() {
@@ -541,100 +520,6 @@ class TseServiceTest {
 
             assertThat(tseService.formatViewApplication(caseData, AUTH_TOKEN, false)).isEqualTo(expected);
         }
-    }
-
-    @ParameterizedTest
-    @MethodSource("selectedApplicationList")
-    void validateGiveDetails_Blank_ReturnErrorMsg(String selectedApplication) {
-        CaseData caseData = CaseDataBuilder.builder().build();
-        caseData.setResTseSelectApplication(selectedApplication);
-        List<String> errors = tseService.validateGiveDetails(caseData);
-        MatcherAssert.assertThat(errors.size(), is(1));
-        MatcherAssert.assertThat(errors.get(0), is(GIVE_DETAIL_MISSING));
-    }
-
-    @ParameterizedTest
-    @MethodSource("selectedApplicationList")
-    void validateGiveDetails_HasDoc_NoErrorMsg(String selectedApplication) {
-        CaseData caseData = CaseDataBuilder.builder().build();
-        caseData.setResTseSelectApplication(selectedApplication);
-        setDocForSelectedApplication(caseData);
-        List<String> errors = tseService.validateGiveDetails(caseData);
-        MatcherAssert.assertThat(errors.size(), is(0));
-    }
-
-    @ParameterizedTest
-    @MethodSource("selectedApplicationList")
-    void validateGiveDetails_HasTextBox_NoErrorMsg(String selectedApplication) {
-        CaseData caseData = CaseDataBuilder.builder().build();
-        caseData.setResTseSelectApplication(selectedApplication);
-        setTextBoxForSelectedApplication(caseData);
-        List<String> errors = tseService.validateGiveDetails(caseData);
-        MatcherAssert.assertThat(errors.size(), is(0));
-    }
-
-    private void setTextBoxForSelectedApplication(CaseData caseData) {
-        switch (caseData.getResTseSelectApplication()) {
-            case TSE_APP_AMEND_RESPONSE -> caseData.setResTseTextBox1("Not Blank");
-            case TSE_APP_CHANGE_PERSONAL_DETAILS -> caseData.setResTseTextBox2("Not Blank");
-            case TSE_APP_CLAIMANT_NOT_COMPLIED -> caseData.setResTseTextBox3("Not Blank");
-            case TSE_APP_CONSIDER_A_DECISION_AFRESH -> caseData.setResTseTextBox4("Not Blank");
-            case TSE_APP_CONTACT_THE_TRIBUNAL -> caseData.setResTseTextBox5("Not Blank");
-            case TSE_APP_ORDER_OTHER_PARTY -> caseData.setResTseTextBox6("Not Blank");
-            case TSE_APP_ORDER_A_WITNESS_TO_ATTEND_TO_GIVE_EVIDENCE -> caseData.setResTseTextBox7("Not Blank");
-            case TSE_APP_POSTPONE_A_HEARING -> caseData.setResTseTextBox8("Not Blank");
-            case TSE_APP_RECONSIDER_JUDGEMENT -> caseData.setResTseTextBox9("Not Blank");
-            case TSE_APP_RESTRICT_PUBLICITY -> caseData.setResTseTextBox10("Not Blank");
-            case TSE_APP_STRIKE_OUT_ALL_OR_PART_OF_A_CLAIM -> caseData.setResTseTextBox11("Not Blank");
-            case TSE_APP_VARY_OR_REVOKE_AN_ORDER -> caseData.setResTseTextBox12("Not Blank");
-            default -> {
-            }
-        }
-    }
-
-    private static Stream<Arguments> selectedApplicationList() {
-        return Stream.of(
-                Arguments.of(TSE_APP_AMEND_RESPONSE),
-                Arguments.of(TSE_APP_CHANGE_PERSONAL_DETAILS),
-                Arguments.of(TSE_APP_CLAIMANT_NOT_COMPLIED),
-                Arguments.of(TSE_APP_CONSIDER_A_DECISION_AFRESH),
-                Arguments.of(TSE_APP_CONTACT_THE_TRIBUNAL),
-                Arguments.of(TSE_APP_ORDER_OTHER_PARTY),
-                Arguments.of(TSE_APP_ORDER_A_WITNESS_TO_ATTEND_TO_GIVE_EVIDENCE),
-                Arguments.of(TSE_APP_POSTPONE_A_HEARING),
-                Arguments.of(TSE_APP_RECONSIDER_JUDGEMENT),
-                Arguments.of(TSE_APP_RESTRICT_PUBLICITY),
-                Arguments.of(TSE_APP_STRIKE_OUT_ALL_OR_PART_OF_A_CLAIM),
-                Arguments.of(TSE_APP_VARY_OR_REVOKE_AN_ORDER));
-    }
-
-    private void setDocForSelectedApplication(CaseData caseData) {
-        switch (caseData.getResTseSelectApplication()) {
-            case TSE_APP_AMEND_RESPONSE -> caseData.setResTseDocument1(createDocumentType());
-            case TSE_APP_CHANGE_PERSONAL_DETAILS -> caseData.setResTseDocument2(createDocumentType());
-            case TSE_APP_CLAIMANT_NOT_COMPLIED -> caseData.setResTseDocument3(createDocumentType());
-            case TSE_APP_CONSIDER_A_DECISION_AFRESH -> caseData.setResTseDocument4(createDocumentType());
-            case TSE_APP_CONTACT_THE_TRIBUNAL -> caseData.setResTseDocument5(createDocumentType());
-            case TSE_APP_ORDER_OTHER_PARTY -> caseData.setResTseDocument6(createDocumentType());
-            case TSE_APP_ORDER_A_WITNESS_TO_ATTEND_TO_GIVE_EVIDENCE ->
-                    caseData.setResTseDocument7(createDocumentType());
-            case TSE_APP_POSTPONE_A_HEARING -> caseData.setResTseDocument8(createDocumentType());
-            case TSE_APP_RECONSIDER_JUDGEMENT -> caseData.setResTseDocument9(createDocumentType());
-            case TSE_APP_RESTRICT_PUBLICITY -> caseData.setResTseDocument10(createDocumentType());
-            case TSE_APP_STRIKE_OUT_ALL_OR_PART_OF_A_CLAIM ->
-                    caseData.setResTseDocument11(createDocumentType());
-            case TSE_APP_VARY_OR_REVOKE_AN_ORDER -> caseData.setResTseDocument12(createDocumentType());
-            default -> {
-            }
-        }
-    }
-
-    private UploadedDocumentType createDocumentType() {
-        UploadedDocumentType uploadedDocumentType = new UploadedDocumentType();
-        uploadedDocumentType.setDocumentBinaryUrl("binaryUrl/documents/");
-        uploadedDocumentType.setDocumentFilename("testFileName");
-        uploadedDocumentType.setDocumentUrl("documentUrl");
-        return uploadedDocumentType;
     }
 
     private CaseData setupCaseDataWithAnApplication() {
